type u

+++ libft_util.c
@@ -58,7 +58,7 @@ void ft_putstr(char *s)
 	}
 }
 
-void	ft_putnbr(int n)
+void	ft_putnbr(long long n)
 {
 
 	if (n == -2147483648)
